backend/requirements.txt
fastapi
uvicorn
sqlmodel
python-jose
passlib[bcrypt]
httpx

